delete expired channels bug resolved

+++ Frontend/channelx/src/components/Home/Home.js
@@ -492,24 +492,27 @@ class Home extends Component {
 
         console.log("Inside DeleteExpiredChannel");
 
-        var time;
         var today = new Date(),
             date = (today.getMonth() + 1) + '/' + today.getDate() + '/' + today.getFullYear();
-        time = Moment(today).format('HH:mm:ss').toString();
 
         console.log(date);
-        db.collection("channels").where("channelEndDate", "<", date)
-            .get()
-            .then(snapshot => {
-                snapshot
-                    .docs
-                    .forEach(doc => {
-                        //console.log("Expired Channel id: " + doc.id);
-                        doc.ref.delete();
-                        messagingChannelDeleter.deleteChannel(doc.id);
+        console.log("here ");
 
-                    })
+        db.collection("channels").get().then(ref => {
+            ref.docs.forEach(doc => {
+
+                var channelEndDate = doc.get("channelEndDate");
+                // var channelStartDate = doc.get("channelStartDate");
+
+                var validDate = Moment(date).isSameOrBefore(channelEndDate);
+
+                if(!validDate){
+                    doc.ref.delete();
+                    messagingChannelDeleter.deleteChannel(doc.id);
+                }
+
             });
+        });
 
     }
